Move to GPLv3 from apache2
An experimental attempt to re-assert some of my rights as
creator of code. This should *not* affect any users. Most corporations
already run lots of GPL software (Linux, gcc, etc). This will only
affect folks who want to make changes and distribute those changes to
users. So if you are setting this up for your company, no problem.
Setting it up publicly, with your own changes? *Still* no problem,
since you arne't distributing the KubeSSH code - you are only
allowing users access to it. This does not trigger a distribution
clause - unlike the AGPL.

The *only* time you will need to make the changes public is if,
for example, you make this into a product available for sale
to the general public. Then you are distributing the KubeSSH code,
and your users (who have purchased the product) can request a copy
of the changes you have made.

TL;DR: Nothing changes for you, unless you are planning on packaging
and shipping a product based on KubeSSH to users.

See https://medium.com/@yuvipanda/removing-open-source-developer-from-my-identity-bc189406ea2a
for some more musings on this