Use deterministic pod names
We were using labels + selectors to determine which user
was on which pod. This is extremely useful when you have many
pods per user. However, we do not support that yet. We also need
to make a netwrok call to figure out name of pod - something needed
for port-forward and exec.

So we go back to the JupyterHub tactic of using a predictable
pod name, so we can get it without having to hit the network.

Cleanup and document a little bit of ensure_running while there

Also move server.py to app.py, because I could not be bothered
to do appropriate git surgery